refactor(Entity): 패키지 정리
