+++ src/main/java/persistence/entity/common/EntityMetaCache.java
@@ -1,4 +1,4 @@
-package persistence.entity;
+package persistence.entity.common;
 
 import persistence.sql.model.Table;
