+++ src/main/java/persistence/entity/entitymanager/EntityEntry.java
@@ -1,4 +1,4 @@
-package persistence.entity;
+package persistence.entity.entitymanager;
 
 public class EntityEntry {
