+++ src/main/java/persistence/entity/entitymanager/EntityPersister.java
@@ -1,6 +1,9 @@
-package persistence.entity;
+package persistence.entity.entitymanager;
 
 import database.Database;
+import persistence.entity.common.EntityBinder;
+import persistence.entity.common.EntityId;
+import persistence.entity.common.EntityMetaCache;
 import persistence.sql.dml.DMLQueryBuilder;
 import persistence.sql.model.Table;
