+++ src/main/java/persistence/entity/entitymanager/Status.java
@@ -1,4 +1,4 @@
-package persistence.entity;
+package persistence.entity.entitymanager;
 
 public enum Status {
     MANAGED,
