+++ src/test/java/persistence/entity/SimpleEntityManagerTest.java
@@ -10,6 +10,11 @@ import org.junit.jupiter.api.*;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import persistence.entity.common.EntityId;
+import persistence.entity.entitymanager.EntityLoader;
+import persistence.entity.entitymanager.EntityManager;
+import persistence.entity.entitymanager.EntityPersister;
+import persistence.entity.entitymanager.SimpleEntityManager;
 import persistence.sql.ddl.DDLQueryBuilder;
 import persistence.sql.dialect.Dialect;
 import persistence.sql.dialect.H2Dialect;
