+++ src/test/java/persistence/sql/QueryTest.java
@@ -4,7 +4,7 @@ import database.DatabaseServer;
 import database.H2;
 import jdbc.JdbcTemplate;
 import org.junit.jupiter.api.*;
-import persistence.entity.EntityId;
+import persistence.entity.common.EntityId;
 import persistence.sql.ddl.DDLQueryBuilder;
 import persistence.sql.dialect.Dialect;
 import persistence.sql.dialect.H2Dialect;
