+++ src/test/java/persistence/sql/dml/DeleteQueryBuilderTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import persistence.entity.EntityBinder;
-import persistence.entity.EntityId;
+import persistence.entity.common.EntityBinder;
+import persistence.entity.common.EntityId;
 import persistence.sql.model.PKColumn;
 import persistence.sql.model.Table;
 import persistence.study.sql.ddl.Person1;
