+++ src/test/java/persistence/study/CustomJpaRepositoryTest.java
@@ -6,7 +6,11 @@ import database.H2;
 import database.SimpleDatabase;
 import jdbc.JdbcTemplate;
 import org.junit.jupiter.api.*;
-import persistence.entity.*;
+import persistence.entity.common.EntityId;
+import persistence.entity.entitymanager.EntityLoader;
+import persistence.entity.entitymanager.EntityManager;
+import persistence.entity.entitymanager.EntityPersister;
+import persistence.entity.entitymanager.SimpleEntityManager;
 import persistence.sql.ddl.DDLQueryBuilder;
 import persistence.sql.dialect.Dialect;
 import persistence.sql.dialect.H2Dialect;
